refactor: ChatMessage, ReadState 도메인 합체 #13
