+++ src/main/java/com/example/chatserver/domain/chatMessage/controller/WebSocketController.java
@@ -6,9 +6,9 @@ import com.example.chatserver.domain.chatMessage.dto.payload.GetMessagePayload;
 import com.example.chatserver.domain.chatMessage.dto.payload.SendMessagePayload;
 import com.example.chatserver.domain.chatMessage.service.ChatMessageService;
 import com.example.chatserver.domain.chatRoom.repository.ChatRoomRepository;
-import com.example.chatserver.domain.readState.dto.payload.ReadStatePayload;
-import com.example.chatserver.domain.readState.dto.request.MarkReadRequest;
-import com.example.chatserver.domain.readState.service.ReadStateService;
+import com.example.chatserver.domain.chatMessage.dto.payload.ReadStatePayload;
+import com.example.chatserver.domain.chatMessage.dto.request.MarkReadRequest;
+import com.example.chatserver.domain.chatMessage.service.ReadStateService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
